requirements.txt
# LUNA AI Agent Dependencies

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0

# Configuration
pyyaml>=6.0

# GUI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Voice (optional)
pyttsx3>=2.90
SpeechRecognition>=3.10.0

# Utilities
python-dotenv>=1.0.0
